Create automatic cisco interface shutdown
This method is loaded directly on the router. I have this cron  job listed for review purposes, only. It's best to use Python and run it from outside the router, and inside the datacenter. That way, if the datacenter is infected with ransomware, the task will not complete, thereby preserving you disaster recovery environment.